src/test/java/leetcode/editor/cn/sort/HeapSortTest.java
package leetcode.editor.cn.sort;

import java.util.Arrays;
import leetcode.editor.cn.SortUtils;
import leetcode.editor.cn.sort.HeapSort.Solution;
import org.junit.Assert;
import org.junit.Test;

public class HeapSortTest {

	@Test
	public void test0() {
		Solution solution = new Solution();
		int[] nums = new int[]{7, 6, 5, 2, 3, 4, 1};
		int[] sorted = SortUtils.sortArray(nums);
		int[] result = solution.sort(nums);
		Assert.assertEquals(Arrays.toString(sorted), Arrays.toString(result));
	}

	@Test
	public void test() {
		Solution solution = new Solution();
		int[] nums = SortUtils.getRandomArray();
		int[] sorted = SortUtils.sortArray(nums);
		int[] result = solution.sort(nums);
		Assert.assertEquals(Arrays.toString(sorted), Arrays.toString(result));
	}

}